File restructure by providers for add_cloud_metadata
