+++ libbeat/docs/processors-config.asciidoc
@@ -343,7 +343,7 @@ _GCE_
 }
 -------------------------------------------------------------------------------
 
-_Tencent Clound_
+_Tencent Cloud_
 
 [source,json]
 -------------------------------------------------------------------------------
