+++ libbeat/processors/add_cloud_metadata/provider_digital_ocean.go
@@ -1,0 +1,22 @@
+package actions
+
+import (
+	"github.com/elastic/beats/libbeat/common"
+	s "github.com/elastic/beats/metricbeat/schema"
+	c "github.com/elastic/beats/metricbeat/schema/mapstriface"
+)
+
+// DigitalOcean Metadata Service
+func newDoMetadataFetcher(config common.Config) (*metadataFetcher, error) {
+	doSchema := func(m map[string]interface{}) common.MapStr {
+		out, _ := s.Schema{
+			"instance_id": c.StrFromNum("droplet_id"),
+			"region":      c.Str("region"),
+		}.Apply(m)
+		return out
+	}
+	doMetadataURI := "/metadata/v1.json"
+
+	fetcher, err := newMetadataFetcher(config, "digitalocean", nil, metadataHost, doSchema, doMetadataURI)
+	return fetcher, err
+}
