+++ libbeat/processors/add_cloud_metadata/provider_tencent_cloud.go
@@ -1,0 +1,40 @@
+package actions
+
+import "github.com/elastic/beats/libbeat/common"
+
+// Tenccent Cloud Metadata Service
+func newQcloudMetadataFetcher(c common.Config) (*metadataFetcher, error) {
+	qcloudMetadataHost := "metadata.tencentyun.com"
+	qcloudMetadataInstanceIDURI := "/meta-data/instance-id"
+	qcloudMetadataRegionURI := "/meta-data/placement/region"
+	qcloudMetadataZoneURI := "/meta-data/placement/zone"
+
+	qcloudSchema := func(m map[string]interface{}) common.MapStr {
+		return common.MapStr(m)
+	}
+
+	urls, err := getMetadataURLs(c, qcloudMetadataHost, []string{
+		qcloudMetadataInstanceIDURI,
+		qcloudMetadataRegionURI,
+		qcloudMetadataZoneURI,
+	})
+	if err != nil {
+		return nil, err
+	}
+	responseHandlers := map[string]responseHandler{
+		urls[0]: func(all []byte, result *result) error {
+			result.metadata["instance_id"] = string(all)
+			return nil
+		},
+		urls[1]: func(all []byte, result *result) error {
+			result.metadata["region"] = string(all)
+			return nil
+		},
+		urls[2]: func(all []byte, result *result) error {
+			result.metadata["availability_zone"] = string(all)
+			return nil
+		},
+	}
+	fetcher := &metadataFetcher{"qcloud", nil, responseHandlers, qcloudSchema}
+	return fetcher, nil
+}
